move encoder implementation to module extensions
